ai-pr-reviewer/requirements.txt
google-genai==0.3.0
requests==2.31.0
PyYAML==6.0.1